Split out delete confirmation modal, and fix associated tests
