+++ packages/components/src/entities/EntityTypeDeleteConfirmModal.spec.tsx
@@ -66,6 +66,7 @@ describe('EntityTypeDeleteConfirmModal', () => {
                 deleteConfirmationActionName="deleteSampleTypes"
                 onCancel={onCancelFn}
                 onConfirm={onConfirmFn}
+                showDeleteComment
             />
         );
